+++ activerecord/test/schema/schema.rb
@@ -105,6 +105,9 @@ ActiveRecord::Schema.define do
     t.column :illustrator_visibility, :integer, default: 0
     t.column :font_size, :integer, default: 0
     t.column :difficulty, :integer, default: 0
+    t.column :privacy, :integer, default: 0
+    t.column :color, :integer, default: 0
+    t.column :font_weight, :integer, default: 0
     t.column :cover, :string, default: "hard"
   end